pr: changed addItem to accept a table of augments properly
This matches the "old way" of passing a table. Where the table
could look like
{
   [100] = 1,
   [50] = 2,
}

where the index is the augid and the value is the augval

The code (as it was converted) was looking at something like
{
   [1] = {100, 1}
}
which is not how this "pass augments to addItem via table" used to work.