Update Trust Adelheid to use Dark Arts / Addendum: Black
Adds Gambits to use Dark Arts and Addendum: Black.
Updates Effects for Dark Arts and Addendum: Black to account for non-PCs.

Notes:

charutils (called by Dark Arts and Addendum: Black), does not work with Trusts and will cause a server crash when called by a Trust currently.
This change creates checks to bypass `recalculateAbilitiesTable()` and `recalculateSkillsTable()` when called by a Trust in the above mentioned scripts.
Instead of applying skill via MODs, applies an amount of MACC equal to (Skill B+ - Skill D) since Skill MODs are also applied via `charutils` and not a viable workaround.
Dark Arts Gambit checks for Addendum: Black so as to not re-apply Dark Arts and create a repeating loop.

Left comments and TODOs.